Move roles from private image to base
These were executed needlessly every time we provision
the private image. If they are moved to the base image
then the provision of the private image will be faster
and the public image will be better. The only problem
is that we now need to update the base image whenever
we change any of these roles.